+++ lib/lma_sdk/lma_sdk/_core/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2025 Amazon.com
+# This file is licensed under the MIT License.
+"""Core implementation modules for LMA SDK."""